Moved some classes around to fit into namespaces better
