+++ LoxFramework/LoxFramework.csproj
@@ -52,9 +52,9 @@
     <Compile Include="Evaluating\AstInterpreter.cs" />
     <Compile Include="AST\Expression.cs" />
     <Compile Include="AST\Statement.cs" />
-    <Compile Include="Environment.cs" />
-    <Compile Include="LoxRunTimeException.cs" />
-    <Compile Include="Extensions.cs" />
+    <Compile Include="Evaluating\Environment.cs" />
+    <Compile Include="Evaluating\LoxRunTimeException.cs" />
+    <Compile Include="Scanning\Extensions.cs" />
     <Compile Include="Interpreter.cs" />
     <Compile Include="InterpreterEventArgs.cs" />
     <Compile Include="Parsing\Parser.cs" />
@@ -66,8 +66,6 @@
   <ItemGroup>
     <None Include="Parsing\Grammer.md" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Utility\" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
